% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm-add-keys-block.R
\name{block_output.dm_add_keys_block}
\alias{block_output.dm_add_keys_block}
\title{Custom output for dm add keys blocks}
\usage{
block_output.dm_add_keys_block(x, result, session)
}
\arguments{
\item{x}{The block object}

\item{result}{The dm result}

\item{session}{Shiny session}
}
\description{
Custom output for dm add keys blocks
}
