% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm-filter-block.R
\name{block_output.dm_filter_block}
\alias{block_output.dm_filter_block}
\title{Custom output for dm filter blocks}
\usage{
block_output.dm_filter_block(x, result, session)
}
\arguments{
\item{x}{The block object}

\item{result}{The dm result}

\item{session}{Shiny session}
}
\description{
Custom output for dm filter blocks
}
