% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm-read-block.R
\name{block_output.dm_read_block}
\alias{block_output.dm_read_block}
\title{Custom output for dm_read_block}
\usage{
\method{block_output}{dm_read_block}(x, result, session)
}
\arguments{
\item{x}{The block object}

\item{result}{The dm result}

\item{session}{Shiny session}
}
\description{
Custom output for dm_read_block
}
