% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm-block.R
\name{block_render_trigger.dm_block}
\alias{block_render_trigger.dm_block}
\title{Custom render trigger for dm blocks}
\usage{
\method{block_render_trigger}{dm_block}(x, session = blockr.core::get_session())
}
\arguments{
\item{x}{The block object}

\item{session}{Shiny session}
}
\description{
Override transform_block's render trigger which tries to access
board options that may not be set. dm blocks don't need pagination.
}
