% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm-add-keys-block.R, R/dm-block.R
\name{block_ui.dm_add_keys_block}
\alias{block_ui.dm_add_keys_block}
\alias{block_ui.dm_block}
\title{Custom UI for dm blocks}
\usage{
\method{block_ui}{dm_add_keys_block}(id, x, ...)

\method{block_ui}{dm_block}(id, x, ...)
}
\arguments{
\item{id}{Namespace ID}

\item{x}{The block object}

\item{...}{Additional arguments}
}
\description{
Custom UI for dm blocks
}
