% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm-example-block.R
\name{dm_example_choices}
\alias{dm_example_choices}
\title{List available dm example datasets}
\usage{
dm_example_choices()
}
\value{
Named character vector.
}
\description{
Returns a named character vector (id -> display name) of dm example datasets
whose required packages are installed.
}
\keyword{internal}
