% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm-example-block.R
\name{dm_example_expr}
\alias{dm_example_expr}
\title{Get a quoted expression for a dm example}
\usage{
dm_example_expr(id)
}
\arguments{
\item{id}{Character, the dataset ID.}
}
\value{
A quoted R expression that evaluates to a \code{dm} object.
}
\description{
Get a quoted expression for a dm example
}
\keyword{internal}
