% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm-block.R
\name{dm_highlight_js}
\alias{dm_highlight_js}
\title{JS for dm diagram node highlight on click}
\usage{
dm_highlight_js(diagram_id)
}
\arguments{
\item{diagram_id}{DOM id of the grViz widget}
}
\description{
JS for dm diagram node highlight on click
}
\keyword{internal}
