% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm-keylines.R
\name{dm_keylines_css}
\alias{dm_keylines_css}
\title{CSS for the Key-lines schematic + table preview}
\usage{
dm_keylines_css()
}
\description{
CSS for the Key-lines schematic + table preview
}
\keyword{internal}
