% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm-keylines.R
\name{dm_keylines_js}
\alias{dm_keylines_js}
\title{JS for the Key-lines schematic: selection, key-cap isolation + click}
\usage{
dm_keylines_js(root_id, input_id)
}
\arguments{
\item{root_id}{DOM id of the schematic root}

\item{input_id}{Fully-namespaced Shiny input id to receive the click}
}
\description{
JS for the Key-lines schematic: selection, key-cap isolation + click
}
\keyword{internal}
