% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm-keylines.R
\name{dm_keylines_memberships}
\alias{dm_keylines_memberships}
\title{Line memberships of one table (for its relationship tags)}
\usage{
dm_keylines_memberships(lines, table_id)
}
\description{
Line memberships of one table (for its relationship tags)
}
\keyword{internal}
