% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm-block.R
\name{new_dm_block}
\alias{new_dm_block}
\title{Create dm Block Constructor}
\usage{
new_dm_block(...)
}
\arguments{
\item{...}{Forwarded to \code{\link[blockr.core:new_transform_block]{blockr.core::new_transform_block()}}}
}
\value{
A block object for creating dm objects
}
\description{
This block combines multiple data frames into a dm (data model) object.
}
\examples{
# Create a dm block
new_dm_block()

if (interactive()) {
  library(blockr.core)
  library(blockr.dm)

  # Combine multiple tables into a dm
  serve(
    new_board(
      blocks = list(
        airlines = new_dataset_block(dataset = "airlines", package = "nycflights13"),
        flights = new_dataset_block(dataset = "flights", package = "nycflights13"),
        dm_obj = new_dm_block()
      ),
      links = c(
        new_link("airlines", "dm_obj", "1"),
        new_link("flights", "dm_obj", "2")
      )
    )
  )
}

}
