% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm-nested-view-block.R
\name{new_dm_nested_view_block}
\alias{new_dm_nested_view_block}
\title{Create dm Nested View Block}
\usage{
new_dm_nested_view_block(root_table = character(), ...)
}
\arguments{
\item{root_table}{Character, the name of the root/parent table to display.
Child tables with foreign keys pointing to this table will be shown
as expandable nested content.}

\item{...}{Forwarded to \code{\link[blockr.core:new_transform_block]{blockr.core::new_transform_block()}}}
}
\value{
A block object for nested dm visualization
}
\description{
Displays a dm object as an interactive nested table where clicking
a row in the parent table expands to show related child tables.
}
\examples{
# Display ADSL with nested ADAE and ADLB
new_dm_nested_view_block(root_table = "adsl")

}
