% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm-pluck-block.R
\name{new_dm_pluck_block}
\alias{new_dm_pluck_block}
\title{dm Pluck Block Constructor}
\usage{
new_dm_pluck_block(table = "", ...)
}
\arguments{
\item{table}{Character, the name of the table to extract. Default "".}

\item{...}{Forwarded to \code{\link[blockr.core:new_transform_block]{blockr.core::new_transform_block()}}}
}
\value{
A block object for extracting tables from dm objects
}
\description{
This block extracts a single table from a dm object as a regular data frame.
Use this after filtering to get a specific table for further analysis
or visualization.
}
\examples{
# Create a dm pluck block
new_dm_pluck_block(table = "flights")

}
