% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm-block.R
\name{block_output.dm_block}
\alias{block_output.dm_block}
\title{Custom output for dm blocks}
\usage{
block_output.dm_block(x, result, session)
}
\arguments{
\item{x}{The block object}

\item{result}{The dm result}

\item{session}{Shiny session}
}
\description{
Displays dm structure instead of trying to render as a table.
}
