% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm-read-block.R
\name{block_ui.dm_read_block}
\alias{block_ui.dm_read_block}
\title{Custom UI for dm_read_block}
\usage{
\method{block_ui}{dm_read_block}(id, x, ...)
}
\arguments{
\item{id}{Namespace ID}

\item{x}{The block object}

\item{...}{Additional arguments}
}
\description{
Custom UI for dm_read_block
}
